feat: add renew token functionality
- verify funciton update
- clean up project
- make reusable token generator

+++ options/validity.go
@@ -2,14 +2,14 @@ package options
 
 type Validity struct {
 	AccessExpiresIn  int64
-	RefreshExpiresIn int64
+	RefreshExpiresIn int
 }
 
 func (v *Validity) GetAccessExpiresIn() int64 {
 	return v.AccessExpiresIn
 }
 
-func (v *Validity) GetRefreshExpiresIn() int64 {
+func (v *Validity) GetRefreshExpiresIn() int {
 	return v.RefreshExpiresIn
 }
